Got somewhat sidetracked with lts
Needed to have the memory_manager init done becuase the input was
getting stopped with that. And since I was already doing the input
for the lts I might as well have just moved the stub code into the init
function for the memory_manager and knocked out all of the input reading
in one session.