Got the input working
Files can now be read in via the lts init function. The lts will then
create the necessary memory blocks. Except for Segments. That has yet to
be completed. I want to complete the other parts of the simulation
before I start working on segmenets since it is slightly different than
both paging and VSP. Now we need to start working on the different ticks
for each part of the simulation, starting with the lts tick. It will
look at every process in the list each tick and if it is time for the
process to arrive it will send it off to the memory managers input
queue.